Fixed Linking issue in shared library
edditingExecutables had a demo involving a shared library. Shared
libraries need linker scripts (which include a specific path to look
in). Now, instead of providing a hardcoded path, we get the current
makefile path, and parse that the find our correct location. It should
work when running in any directory now.